Create SmimeKey class and move functions there
